requirements.txt
APScheduler==3.5.3
atomicwrites==1.2.1
attrs==18.2.0
celery==4.2.1
click==6.7
cmake==3.12.0
colorama==0.3.9
dash-core-components==0.30.2
dash-html-components==0.13.2
dash-renderer==0.14.1
dash-table-experiments==0.6.0
dash==0.27.0
Flask-APScheduler==1.10.1
Flask-Compress==1.4.0
Flask-MQTT==1.0.4
holidays==0.9.8
imutils==0.5.1
ipykernel==5.0.0
ipython-genutils==0.2.0
ipython==7.0.1
ipywidgets==7.4.2
jsonlines==1.2.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter==1.0.0
kiwisolver==1.0.1
more-itertools==4.3.0
mysql-connector-python
mysql-connector
notebook==5.7.0
numpy==1.15.1
opencv-python
paho-mqtt==1.4.0
pandas==0.23.4
pathos==0.2.2.1
pluggy==0.7.1
py==1.6.0
pymysql==0.9.2
pytest==3.8.2
python-dateutil==2.7.3
pytz==2018.5
PyUtilib==5.6.3
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.3.0
urllib3==1.22
gunicorn
flask
flask-sqlalchemy
flask_table
Flask-WTF
psycopg2-binary
SpeechRecognition
Flask-MySQLdb
PyYAML
werkzeug
watson_developer_cloud